changes to the inventory composable system, more modular
to whoever is reviewing this, I can only apologize, it is far too encompassing compared to the previous commits.
however, i aim to make it more clear through the power of **markdown**. *(what you're currently reading. it is about 1AM, and i am tired, so indulge me)*.
- changes to Character.kt: made the fields mutable, so that the character clothing can actually be swapped. No idea why i didn't do this from the start.
- changes to Clothing.kt:
- - Inventory now contains a function that handles closing of the menu on all levels. Advantages of this is that 1. it is tied to the state and 2. it is passed through all levels, where closing is needed. By this i mean that closing is necessary on the bottom of the hierarchy when an item is clicked, but also on dialog-level when ondismiss() is called. this function at the top handles all that.
also in Clothing.kt: the rendering of each item is moved to a separate custom Card-component, mainly for more readable code and the built in  onclick.

i understand this commit is too big, please just add a comment to the commit or message me with any questions